Testing changes
Testing changes

+++ UtilityMediaPlayer/src/Player.java
@@ -3,7 +3,7 @@ public interface Player
 {
 	public boolean open(String fileName);
 	public boolean clear();
-	public void volumeChange(int newVolume);
+	public void volumeChange(double newVolume);
 	public void alternatePlayback();
 	public void skipPlayback();
 	public void changePosition(long playbackPosition);
